feat: add human-friendly entry point
* Added /start as entry point, for humans and iPXE

The '/poll/1/${netX/mac:hexhyp}' is fairly iPXE internal, not something
that invites curious people to look better at shoelaces. Neither it is
looking nice in DHCP server configuration.

This change adds a HTTP handler for '/start'. And '/start' points
to '/poll/1/${netX/mac:hexhyp}'. The benefit of it is that human visible
documentation can replace the "voodoo" '/poll/1/${netX/mac:hexhyp}'
with "friendly" '/start'.

Because it is an addition are the existing HTTP handlers not effected,
neither the installed deployments effected.

Signed-off-by: Geert Stappers <stappers@stappers.it>