Added the serve_more utility
`serve_more` is a shell script for creating symbolic links in the web
directory. Those symlinks point to a vmlinuz or a initrd elsewhere on
the server. It is for making shoelaces more webserver for netbooting.

Example, files like

  /usr/lib/debian-installer/images/11/arm64/text/debian-installer/arm64/initrd.gz
  /usr/lib/debian-installer/images/11/arm64/text/debian-installer/arm64/linux

on the shoelaces server can be served as

  http://shoelaces.lan:8081/static/di/arm64/initrd
  http://shoelaces.lan:8081/static/di/arm64/linux

by providing a file with lines like

  di/arm64/initrd  /usr/lib/debian-installer/images/11/arm64/text/debian-installer/arm64/initrd.gz
  di/arm64/linux  /usr/lib/debian-installer/images/11/arm64/text/debian-installer/arm64/linux

to the script.

Signed-off-by: Geert Stappers <stappers@stappers.it>